v1 Functionality (#3)
## Why is this change necessary?
Needed to be able to handle a broader range of event types, as well as
delaying after upload. Also sending heartbeat to cloudwatch. There was
an issue that needed to be fixed where S3 Object tags included invalid
characters.


 ## How is this change tested?
Unit tests and locally on Windows to confirm heartbeat and file upload.