Update instance tag (#9)
## Link to Issue or Slack thread

https://github.com/LabAutomationAndScreening/copier-cloud-courier-infrastructure/pull/7


 ## Why is this change necessary?
Have the agent update the Managed Instance tag with the version of the
upload agent when it boots up.


 ## How does this change address the issue?
It does that.


 ## What side effects does this change have?
None


 ## How is this change tested?
Stubbed unit tests.  And locally on my PC.


 ## Other
Added a `--version` parameter to the CLI. Added a test of the executable
to determine that it works (but that doesn't run in CI yet).